Begin Rework - Renamed class. Moved client to base. Move get_quarter to _index, sanitize_date to filing. rename Filing to CompanyFilings method Filings to return most relevant class moderize CIK lookup check_urls_exist should be a static method, and should not return remove abstract client. automatically select default save method fix #159, add downloading back switch out for temp directory move logic into save_filings to avoid calling get_urls for ultrafast_bulk_save. move options into save_filings switch out for kwargs remove useless method, cleanup async, turn client into property Fix test import errors (still a lot of problems)
