+++ secedgar/filings/__init__.py
@@ -1,5 +1,5 @@
 from secedgar.filings.cik_lookup import CIKLookup  # noqa: F401
 from secedgar.filings.daily import DailyFilings  # noqa:F401
-from secedgar.filings.filing import Filing  # noqa:F401
+from secedgar.filings.company import CompanyFilings  # noqa:F401
 from secedgar.filings.filing_types import FilingType  # noqa:F401
 from secedgar.filings.master import MasterFilings  # noqa:F401
