+++ secedgar/filings/_base.py
@@ -5,39 +5,18 @@ from abc import ABC, abstractmethod
 from secedgar.parser import MetaParser
 
 
-class AbstractFiling(ABC):
+class FilingStrategy(ABC):
     """Abstract base class for all SEC EDGAR filings.
 
     .. versionadded:: 0.1.5
     """
-
-    def extract_meta(self, directory, out_dir=None, create_subdir=True, rm_infile=False):
-        """Extract meta data from filings in directory."""
-        for root, _, files in os.walk(directory):
-            for file in files:
-                if file.endswith('.txt'):
-                    MetaParser().process(os.path.join(root, file),
-                                         out_dir=out_dir,
-                                         create_subdir=create_subdir,
-                                         rm_infile=rm_infile)
+    def __init__(self, **kwargs):
+        self._client = kwargs.get('client', NetworkClient())
 
     @property
-    @abstractmethod
     def client(self):
         """``secedgar.client._base``: Client to use to make requests."""
-        pass  # pragma: no cover
-
-    @property
-    @abstractmethod
-    def params(self):
-        """:obj:`dict`: Parameters to include in requests."""
-        pass  # pragma: no cover
-
-    @property
-    @abstractmethod
-    def path(self):
-        """str: Path added to client base."""
-        pass  # pragma: no cover
+        return self._client
 
     @abstractmethod
     def get_urls(self, **kwargs):
@@ -84,7 +63,19 @@ class AbstractFiling(ABC):
         stripped = "".join(c for c in path if c in allowed)
         return stripped.replace(" ", "_")
 
-    def _check_urls_exist(self):
+    @staticmethod
+    def extract_meta(directory, out_dir=None, create_subdir=True, rm_infile=False):
+        """Extract meta data from filings in directory."""
+        for root, _, files in os.walk(directory):
+            for file in files:
+                if file.endswith('.txt'):
+                    MetaParser().process(os.path.join(root, file),
+                                         out_dir=out_dir,
+                                         create_subdir=create_subdir,
+                                         rm_infile=rm_infile)
+
+    @staticmethod   
+    def check_urls_exist(urls):
         """Wrapper around `get_urls` to check if there is a positive number of URLs.
 
         .. note:: This method will not check if the URLs are valid. Simply if they exist.
@@ -95,7 +86,5 @@ class AbstractFiling(ABC):
         Returns:
             urls (dict): Result of `get_urls` method.
         """
-        urls = self.get_urls()
         if all(len(urls[cik]) == 0 for cik in urls.keys()):
             raise ValueError("No filings available.")
-        return urls
