+++ secedgar/filings/company.py
@@ -4,14 +4,13 @@ import os
 import warnings
 
 from secedgar.client import NetworkClient
-from secedgar.filings._base import AbstractFiling
+from secedgar.filings._base import FilingStrategy
 from secedgar.filings.cik_lookup import CIKLookup
 from secedgar.filings.filing_types import FilingType
-from secedgar.utils import sanitize_date
 from secedgar.utils.exceptions import FilingTypeError
 
 
-class Filing(AbstractFiling):
+class CompanyFilings(FilingStrategy):
     """Base class for receiving EDGAR filings.
 
     Attributes:
@@ -54,21 +53,6 @@ class Filing(AbstractFiling):
         self._client = client if client is not None else NetworkClient(**kwargs)
 
     @property
-    def path(self):
-        """str: Path added to client base."""
-        return "cgi-bin/browse-edgar"
-
-    @property
-    def params(self):
-        """:obj:`dict`: Parameters to include in requests."""
-        return self._params
-
-    @property
-    def client(self):
-        """``secedgar.client._base``: Client to use to make requests."""
-        return self._client
-
-    @property
     def start_date(self):
         """Union([datetime.datetime, str]): Date before which no filings are fetched."""
         return self._start_date
@@ -81,6 +65,28 @@ class Filing(AbstractFiling):
         else:
             self._start_date = None
 
+    @staticmethod
+    def sanitize_date(date):
+        """Sanitizes date to be in acceptable format for EDGAR.
+
+        Args:
+            date (Union[datetime.datetime, str]): Date to be sanitized for request.
+
+        Returns:
+            date (str): Properly formatted date in 'YYYYMMDD' format.
+
+        Raises:
+            TypeError: If date is not in format YYYYMMDD as str or int.
+        """
+        if isinstance(date, datetime.datetime):
+            return date.strftime("%Y%m%d")
+        elif isinstance(date, str):
+            if len(date) != 8:
+                raise TypeError('Date must be of the form YYYYMMDD')
+        elif isinstance(date, int):
+            if date < 10 ** 7 or date > 10 ** 8:
+                raise TypeError('Date must be of the form YYYYMMDD')
+        return date
     @property
     def end_date(self):
         """Union([datetime.datetime, str]): Date after which no filings are fetched."""
@@ -163,14 +169,14 @@ class Filing(AbstractFiling):
             txt_urls (list of str): Up to the desired number of URLs for that specific company
             if available.
         """
-        self.params['CIK'] = cik
+        self._params['CIK'] = cik
         links = []
-        self.params["start"] = 0  # set start back to 0 before paginating
+        self._params["start"] = 0  # set start back to 0 before paginating
 
         while self.count is None or len(links) < self.count:
-            data = self.client.get_soup(self.path, self.params, **kwargs)
+            data = self.client.get_soup("cgi-bin/browse-edgar", self._params, **kwargs)
             links.extend([link.string for link in data.find_all("filinghref")])
-            self.params["start"] += self.client.batch_size
+            self._params["start"] += self.client.batch_size
             if len(data.find_all("filinghref")) == 0:  # no more filings
                 break
 
@@ -202,7 +208,8 @@ class Filing(AbstractFiling):
         Raises:
             ValueError: If no text urls are available for given filing object.
         """
-        urls = self._check_urls_exist()
+        urls = self.get_urls()
+        self.check_urls_exist(urls)
 
         if dir_pattern is None:
             dir_pattern = os.path.join('{cik}', '{type}')
@@ -211,6 +218,7 @@ class Filing(AbstractFiling):
 
         inputs = []
         for cik, links in urls.items():
+            cik = urls
             formatted_dir = dir_pattern.format(cik=cik, type=self.filing_type.value)
             for link in links:
                 formatted_file = file_pattern.format(
