+++ secedgar/tests/filings/test_cik_lookup.py
@@ -5,7 +5,21 @@ from secedgar.client import NetworkClient
 from secedgar.tests.conftest import MockResponse
 import secedgar.utils
 from secedgar.utils.exceptions import EDGARQueryError
+import gzip
 
+class MockCIKMapResponse:
+    def __init__(self, *args, **kwargs):
+        with gzip.open(datapath("utils", "cik_map.json.gz"), 'rt') as f:
+            self.text = f.read()
+
+@pytest.fixture
+def client():
+    return NetworkClient()
+
+
+@pytest.fixture
+def ticker_lookups():
+    return ["AAPL", "FB", "GOOGL", "NFLX", "MSFT"]
 
 @pytest.fixture
 def mock_get_cik_map(monkeypatch):
@@ -40,6 +54,96 @@ def mock_single_cik_not_found(monkeypatch):
 
 
 class TestCIKLookup(object):
+    @pytest.mark.parametrize(
+        "ticker,cik",
+        [
+            ("AAPL", "320193"),
+            ("FB", "1326801"),
+            ("MSFT", "789019")
+        ]
+    )
+    def test_get_cik_map(self, ticker, cik, monkeypatch):
+        monkeypatch.setattr(requests, 'get', MockCIKMapResponse)
+        cik_map = get_cik_map()
+        assert cik_map[ticker] == cik
+
+    @pytest.mark.parametrize(
+        "name,cik",
+        [
+            ("Apple Inc.", "320193"),
+            ("NIKE, Inc.", "320187"),
+            ("MICROSOFT CORP", "789019"),
+        ]
+    )
+    def test_get_company_name_map(self, name, cik, monkeypatch):
+        monkeypatch.setattr(requests, 'get', MockCIKMapResponse)
+        name_map = get_cik_map(key="title")
+        assert name_map[name] == cik
+
+    @pytest.mark.parametrize(
+        "key",
+        [
+            "Ticker",
+            "Title",
+            "CIK",
+            "Company Name"
+        ]
+    )
+    def test_get_cik_map_bad_keys(self, key):
+        with pytest.raises(ValueError):
+            get_cik_map(key=key)
+
+    def test_client_property(self, client, ticker_lookups):
+        validator = _CIKValidator(ticker_lookups, client=client)
+        assert validator.client == client
+
+    @pytest.mark.parametrize(
+        "bad_lookups",
+        [
+            [],
+            "",
+            ["AAPL", 4, "FB"]
+        ]
+    )
+    def test_empty_lookups_raises_type_error(self, bad_lookups):
+        with pytest.raises(TypeError):
+            _CIKValidator(lookups=bad_lookups)
+
+    def test_lookups_property(self, ticker_lookups):
+        validator = _CIKValidator(lookups=ticker_lookups)
+        assert validator.lookups == ticker_lookups
+
+    @pytest.mark.parametrize(
+        "bad_lookup",
+        [
+            "",
+            4
+        ]
+    )
+    def test_validate_lookup(self, bad_lookup):
+        with pytest.raises(TypeError):
+            _CIKValidator._validate_lookup(bad_lookup)
+
+    @pytest.mark.parametrize(
+        "bad_cik",
+        [
+            "1234",
+            1234,
+            1234567890,
+            "AAPL",
+            "",
+            None,
+        ]
+    )
+    def test_validate_cik_on_bad_ciks(self, bad_cik):
+        with pytest.raises(CIKError):
+            _CIKValidator._validate_cik(bad_cik)
+
+    def test_params_reset_after_get_cik(self, ticker_lookups, client):
+        validator = _CIKValidator(lookups=ticker_lookups, client=client)
+        validator._get_cik(ticker_lookups[0])
+        assert validator.params.get("CIK") is None and validator.params.get("company") is None
+
     @pytest.mark.parametrize(
         "lookup,expected",
         [
